Expect path to wallet file, not dir. Use a cache for account addresses.
Closes #78, taking the approach mentioned here:

https://github.com/FuelLabs/forc-wallet/issues/78#issuecomment-1425160146

When a path to a wallet is specified, we now expect a path to a keystore
file. The file can be anywhere on the user's machine and no longer
requires a `.accounts` file in the same directory to track derived
accounts.

Instead, we now track derived accounts within `~/.fuel/wallets/accounts`,
where each wallet has its own unique cache directory named by hashing
the wallet's ciphertext into a fixed-length hex string. Each directory
contains a file for each derived account address, where the file name is
the wallet's account index and the file itself contains the address as a
utf8 string.

This PR also removes the `serialtest` dev dependency in favour of
generating unique directory names for temporary test directories.