Add `--unverified` flag to commands that display cached addresses
Closes #77 as the main concern with storing addresses in plain text is
that it makes them an easy target of corruption.

This PR changes the behaviour of the `forc wallet accounts` and `forc
wallet account <ix>` commands to always require a password by default.
This ensures the printed addresses are valid. This was tricky to
guarantee in the old behaviour as the cached plain-text files could have
been corrupted or maliciously altered since they were last written.

Both commands now accept an `--unverified` flag that emulates the old
behaviour and purely reads the addresses from the cache.