Re-add the `forc wallet sign` subcommand, remove `forc wallet sign-private`
This re-introduces the `forc wallet sign` subcommand. This re-uses the
existing logic for signing via the account subcommand, though provides
the additional option of signing with a private key.

In the future, we can potentially enter an interactive mode in the case
that the user provides neither the `--account` and `--private` flags.