Add derivation for plain address into forc-wallet
By default, we are given the Bech32 format of the public key from the forc-wallet account command, whereas, for Sway, it is required that we use a plain address in the contract. For the convenience of developers, I rigged up the functionality to derive the plain address directly from the CLI with a unit test that sanity checks the type conversion.