Working on `forc wallet list`
Fixes #149

The first step is to cache some addresses by default when a wallet is created,
the second step would be to make `list` as alias of `balance --unverified`.

The `list` account now will request a minium number of accounts to derive and
cache if there are not enough. To derive new accounts the wallet must be
unlocked requesting the password to the user