Improve `list` command
`list` command will request the password to unlock the wallet to derive the requested number of addresses. The derived addresses will be stored in cache.

If `--unverified` is passed and the cache has enough addresses, the cache will be used, otherwise the user will be requested to unlock their wallet.