feat: implement check sum reporting and update default target (#210)
This PR implements couple of things:

1. Adds a new target
2. Adds checksum routines at places where we print user account, and
accept user account (transfer to address).
3. Bumps sdk version to latest, 0.66.8.


### Checksum support

1. At the places where we output hex address we now output checksum
encoded hex address so that copied and pasted into any other wallet,
users will not face any warning.
2. At places where we accept a user account address, (currently only
`transfer`'s `to` field), we query the provider to see if this account
actually belongs to an user (i.e it is not a blob id, contract id etc).

I needed to add an unwrap, at a place where I am sure that the result is
not going to be failing. After this release I will remove it by
rewriting the transfer section of the wallet.


We first check if the checksum is valid while parsing the `To ` field.
If that is not the case that is an immediate error. If it is indeed with
the correct checksum then we proceed to querying the provider to see if
it is an actual address.


### Couple of examples
Here I try to transfer some funds to a contract, which has invalid 


![image](https://github.com/user-attachments/assets/9843f4a6-4020-431a-b217-47ddd5a04817)

Here i try to transfer to my account with valid check sum.


![image](https://github.com/user-attachments/assets/155866d0-4e92-43ad-b7fa-0e4200db8047)